+++ x/starname/account_handlers_test.go
@@ -3,14 +3,13 @@ package starname
 import (
 	"bytes"
 	"errors"
-	"github.com/iov-one/iovns/tutils"
+	"github.com/iov-one/iovns/pkg/utils"
 	"github.com/iov-one/iovns/x/starname/keeper/executor"
 	"reflect"
 	"testing"
 	"time"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/iov-one/iovns"
 	"github.com/iov-one/iovns/x/configuration"
 	"github.com/iov-one/iovns/x/starname/keeper"
 	"github.com/iov-one/iovns/x/starname/types"
@@ -27,14 +26,14 @@ func Test_Close_handlerMsgAddAccountCertificates(t *testing.T) {
 				})
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 					Type:       types.ClosedDomain,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
+					Name:       utils.StrPtr("test"),
 					ValidUntil: 0,
 					Owner:      keeper.AliceKey,
 				}).Create()
@@ -66,14 +65,14 @@ func Test_Open_handlerMsgAddAccountCertificates(t *testing.T) {
 				})
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 					Type:       types.OpenDomain,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
+					Name:       utils.StrPtr("test"),
 					ValidUntil: 0,
 					Owner:      keeper.AliceKey,
 				}).Create()
@@ -137,7 +136,7 @@ func Test_Common_handlerMsgAddAccountCertificates(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				//
@@ -159,14 +158,14 @@ func Test_Common_handlerMsgAddAccountCertificates(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.AliceKey,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -187,14 +186,14 @@ func Test_Common_handlerMsgAddAccountCertificates(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -221,14 +220,14 @@ func Test_Common_handlerMsgAddAccountCertificates(t *testing.T) {
 				})
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("test"),
-					ValidUntil:   iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Name:         utils.StrPtr("test"),
+					ValidUntil:   utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Owner:        keeper.AliceKey,
 					Certificates: []types.Certificate{[]byte("test")},
 				}).Create()
@@ -256,14 +255,14 @@ func Test_Common_handlerMsgAddAccountCertificates(t *testing.T) {
 				})
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("test"),
-					ValidUntil:   iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Name:         utils.StrPtr("test"),
+					ValidUntil:   utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Owner:        keeper.AliceKey,
 					Certificates: nil,
 				}).Create()
@@ -300,14 +299,14 @@ func Test_Common_handlerMsgAddAccountCertificates(t *testing.T) {
 				})
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("test"),
-					ValidUntil:   iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Name:         utils.StrPtr("test"),
+					ValidUntil:   utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Owner:        keeper.AliceKey,
 					Certificates: []types.Certificate{[]byte("1")},
 				}).Create()
@@ -344,14 +343,14 @@ func Test_Common_handlerMsgAddAccountCertificates(t *testing.T) {
 				})
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.AliceKey,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -369,7 +368,7 @@ func Test_Common_handlerMsgAddAccountCertificates(t *testing.T) {
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				expected := []types.Certificate{[]byte("test")}
 				account := new(types.Account)
-				ok := k.AccountStore(ctx).Read((&types.Account{Domain: "test", Name: tutils.StrPtr("test")}).PrimaryKey(), account)
+				ok := k.AccountStore(ctx).Read((&types.Account{Domain: "test", Name: utils.StrPtr("test")}).PrimaryKey(), account)
 				if !ok {
 					t.Fatal("account not found")
 				}
@@ -388,14 +387,14 @@ func Test_Closed_handlerMsgDeleteAccountCertificate(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.AliceKey,
 					Type:       types.ClosedDomain,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("test"),
+					Name:         utils.StrPtr("test"),
 					ValidUntil:   0,
 					Owner:        keeper.AliceKey,
 					Certificates: []types.Certificate{[]byte("test")},
@@ -414,7 +413,7 @@ func Test_Closed_handlerMsgDeleteAccountCertificate(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				ok := k.AccountStore(ctx).Read((&types.Account{Domain: "test", Name: tutils.StrPtr("test")}).PrimaryKey(), account)
+				ok := k.AccountStore(ctx).Read((&types.Account{Domain: "test", Name: utils.StrPtr("test")}).PrimaryKey(), account)
 				if !ok {
 					t.Fatal("account not found")
 				}
@@ -437,14 +436,14 @@ func Test_Open_handlerMsgDeleteAccountCertificate(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.AliceKey,
 					Type:       types.OpenDomain,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("test"),
+					Name:         utils.StrPtr("test"),
 					ValidUntil:   0,
 					Owner:        keeper.AliceKey,
 					Certificates: []types.Certificate{[]byte("test")},
@@ -472,14 +471,14 @@ func Test_Common_handlerMsgDeleteAccountCertificate(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.AliceKey,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -500,14 +499,14 @@ func Test_Common_handlerMsgDeleteAccountCertificate(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.AliceKey,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -528,14 +527,14 @@ func Test_Common_handlerMsgDeleteAccountCertificate(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.AliceKey,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Owner:      keeper.BobKey,
 				}).Create()
 			},
@@ -556,14 +555,14 @@ func Test_Common_handlerMsgDeleteAccountCertificate(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.AliceKey,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -584,14 +583,14 @@ func Test_Common_handlerMsgDeleteAccountCertificate(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Admin:      keeper.AliceKey,
 				}).Create()
 				// add mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("test"),
-					ValidUntil:   iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Name:         utils.StrPtr("test"),
+					ValidUntil:   utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Owner:        keeper.AliceKey,
 					Certificates: []types.Certificate{[]byte("test")},
 				}).Create()
@@ -610,7 +609,7 @@ func Test_Common_handlerMsgDeleteAccountCertificate(t *testing.T) {
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				// check if certificate is still present
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					t.Fatal("account not found")
 				}
@@ -639,7 +638,7 @@ func Test_Closed_handlerMsgDeleteAccount(t *testing.T) {
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain: "test",
-					Name:   tutils.StrPtr("test"),
+					Name:   utils.StrPtr("test"),
 					Owner:  keeper.BobKey,
 				}).Create()
 			},
@@ -655,7 +654,7 @@ func Test_Closed_handlerMsgDeleteAccount(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if exists {
 					t.Fatalf("handlerMsgDeleteAccount() account was not deleted")
 				}
@@ -672,7 +671,7 @@ func Test_Closed_handlerMsgDeleteAccount(t *testing.T) {
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain: "test",
-					Name:   tutils.StrPtr("test"),
+					Name:   utils.StrPtr("test"),
 					Owner:  keeper.BobKey,
 				}).Create()
 			},
@@ -698,7 +697,7 @@ func Test_Closed_handlerMsgDeleteAccount(t *testing.T) {
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain: "test",
-					Name:   tutils.StrPtr("test"),
+					Name:   utils.StrPtr("test"),
 					Owner:  keeper.AliceKey,
 				}).Create()
 			},
@@ -736,7 +735,7 @@ func Test_Open_handlerMsgDeleteAccount(t *testing.T) {
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain: "test",
-					Name:   tutils.StrPtr("test"),
+					Name:   utils.StrPtr("test"),
 					Owner:  keeper.BobKey,
 				}).Create()
 			},
@@ -769,7 +768,7 @@ func Test_Open_handlerMsgDeleteAccount(t *testing.T) {
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain: "test",
-					Name:   tutils.StrPtr("test"),
+					Name:   utils.StrPtr("test"),
 					Owner:  keeper.BobKey,
 				}).Create()
 			},
@@ -786,7 +785,7 @@ func Test_Open_handlerMsgDeleteAccount(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if exists {
 					t.Fatalf("handlerMsgDeleteAccount() account was not deleted")
 				}
@@ -809,7 +808,7 @@ func Test_Open_handlerMsgDeleteAccount(t *testing.T) {
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain: "test",
-					Name:   tutils.StrPtr("test"),
+					Name:   utils.StrPtr("test"),
 					Owner:  keeper.AliceKey,
 				}).Create()
 			},
@@ -845,7 +844,7 @@ func Test_Open_handlerMsgDeleteAccount(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if exists {
 					t.Fatalf("handlerMsgDeleteAccount() account was not deleted")
 				}
@@ -868,7 +867,7 @@ func Test_Open_handlerMsgDeleteAccount(t *testing.T) {
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain: "test",
-					Name:   tutils.StrPtr("test"),
+					Name:   utils.StrPtr("test"),
 					Owner:  keeper.AliceKey,
 				}).Create()
 			},
@@ -886,7 +885,7 @@ func Test_Open_handlerMsgDeleteAccount(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if exists {
 					t.Fatalf("handlerMsgDeleteAccount() account was not deleted")
 				}
@@ -909,7 +908,7 @@ func Test_Open_handlerMsgDeleteAccount(t *testing.T) {
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain: "test",
-					Name:   tutils.StrPtr("test"),
+					Name:   utils.StrPtr("test"),
 					Owner:  keeper.AliceKey,
 				}).Create()
 			},
@@ -927,7 +926,7 @@ func Test_Open_handlerMsgDeleteAccount(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if exists {
 					t.Fatalf("handlerMsgDeleteAccount() account was not deleted")
 				}
@@ -983,7 +982,7 @@ func Test_Common_handlerMsgDeleteAccount(t *testing.T) {
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain: "test",
-					Name:   tutils.StrPtr("test"),
+					Name:   utils.StrPtr("test"),
 					Owner:  keeper.BobKey,
 				}).Create()
 			},
@@ -999,7 +998,7 @@ func Test_Common_handlerMsgDeleteAccount(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if exists {
 					t.Fatalf("handlerMsgDeleteAccount() account was not deleted")
 				}
@@ -1013,7 +1012,7 @@ func Test_Common_handlerMsgDeleteAccount(t *testing.T) {
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain: "test",
-					Name:   tutils.StrPtr("test"),
+					Name:   utils.StrPtr("test"),
 					Owner:  keeper.BobKey,
 				}).Create()
 			},
@@ -1029,7 +1028,7 @@ func Test_Common_handlerMsgDeleteAccount(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if exists {
 					t.Fatalf("handlerMsgDeleteAccount() account was not deleted")
 				}
@@ -1112,7 +1111,7 @@ func Test_ClosedDomain_handlerMsgRegisterAccount(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					t.Fatal("account test not found")
 				}
@@ -1198,11 +1197,11 @@ func Test_OpenDomain_handleMsgRegisterAccount(t *testing.T) {
 			},
 			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					t.Fatal("account test not found")
 				}
-				expected := iovns.TimeToSeconds(time.Unix(11, 0))
+				expected := utils.TimeToSeconds(time.Unix(11, 0))
 				if account.ValidUntil != expected {
 					t.Fatalf("unexpected account valid until %d, expected %d", account.ValidUntil, expected)
 				}
@@ -1467,7 +1466,7 @@ func Test_Common_handleMsgRegisterAccount(t *testing.T) {
 				// add an account that we are gonna try to overwrite
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("exists"),
+					Name:         utils.StrPtr("exists"),
 					Owner:        keeper.AliceKey,
 					ValidUntil:   0,
 					Resources:    nil,
@@ -1558,8 +1557,8 @@ func Test_Closed_handlerMsgRenewAccount(t *testing.T) {
 				// set mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(time.Unix(1000, 0)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(time.Unix(1000, 0)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -1641,8 +1640,8 @@ func Test_Open_handlerMsgRenewAccount(t *testing.T) {
 				// set mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(time.Unix(1, 0)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(time.Unix(1, 0)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -1657,12 +1656,12 @@ func Test_Open_handlerMsgRenewAccount(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					t.Fatal("account not found")
 				}
 				want := ctx.BlockTime().Add(k.ConfigurationKeeper.GetConfiguration(ctx).AccountRenewalPeriod)
-				if account.ValidUntil != iovns.TimeToSeconds(want) {
+				if account.ValidUntil != utils.TimeToSeconds(want) {
 					t.Fatalf("handlerMsgRenewAccount() want: %d, got: %d", want.Unix(), account.ValidUntil)
 				}
 			},
@@ -1687,8 +1686,8 @@ func Test_Open_handlerMsgRenewAccount(t *testing.T) {
 				// set mock account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(time.Unix(1, 0)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(time.Unix(1, 0)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -1709,7 +1708,7 @@ func Test_Open_handlerMsgRenewAccount(t *testing.T) {
 					t.Fatal("domain not found")
 				}
 				account := new(types.Account)
-				exists = k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists = k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					t.Fatal("account not found")
 				}
@@ -1737,14 +1736,14 @@ func Test_Closed_handlerMsgReplaceAccountResources(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 					Type:       types.ClosedDomain,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
+					Name:       utils.StrPtr("test"),
 					ValidUntil: 0,
 					Owner:      keeper.AliceKey,
 				}).Create()
@@ -1785,14 +1784,14 @@ func Test_Open_handlerMsgReplaceAccountResources(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 					Type:       types.OpenDomain,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
+					Name:       utils.StrPtr("test"),
 					ValidUntil: 0,
 					Owner:      keeper.AliceKey,
 				}).Create()
@@ -1832,14 +1831,14 @@ func Test_Common_handlerMsgReplaceAccountResources(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -1872,14 +1871,14 @@ func Test_Common_handlerMsgReplaceAccountResources(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -2009,7 +2008,7 @@ func Test_Common_handlerMsgReplaceAccountResources(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 			},
@@ -2042,14 +2041,14 @@ func Test_Common_handlerMsgReplaceAccountResources(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -2083,14 +2082,14 @@ func Test_Common_handlerMsgReplaceAccountResources(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -2116,7 +2115,7 @@ func Test_Common_handlerMsgReplaceAccountResources(t *testing.T) {
 					Resource: "valid",
 				}}
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					t.Fatal("account not found")
 				}
@@ -2141,14 +2140,14 @@ func Test_Closed_handlerMsgReplaceAccountMetadata(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 					Type:       types.ClosedDomain,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
+					Name:       utils.StrPtr("test"),
 					ValidUntil: 0,
 					Owner:      keeper.AliceKey,
 				}).Create()
@@ -2180,14 +2179,14 @@ func Test_Open_handlerMsgReplaceAccountMetadata(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 					Type:       types.OpenDomain,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
+					Name:       utils.StrPtr("test"),
 					ValidUntil: 0,
 					Owner:      keeper.AliceKey,
 				}).Create()
@@ -2251,7 +2250,7 @@ func Test_Common_handlerMsgReplaceAccountMetadata(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 			},
@@ -2272,14 +2271,14 @@ func Test_Common_handlerMsgReplaceAccountMetadata(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -2300,14 +2299,14 @@ func Test_Common_handlerMsgReplaceAccountMetadata(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -2332,14 +2331,14 @@ func Test_Common_handlerMsgReplaceAccountMetadata(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -2373,14 +2372,14 @@ func Test_Common_handlerMsgReplaceAccountMetadata(t *testing.T) {
 				// create domain
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Admin:      keeper.BobKey,
 				}).Create()
 				// create account
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
-					ValidUntil: iovns.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
+					Name:       utils.StrPtr("test"),
+					ValidUntil: utils.TimeToSeconds(time.Now().Add(1000 * time.Hour)),
 					Owner:      keeper.AliceKey,
 				}).Create()
 			},
@@ -2398,7 +2397,7 @@ func Test_Common_handlerMsgReplaceAccountMetadata(t *testing.T) {
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				expected := "https://test.com"
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					t.Fatal("account not found")
 				}
@@ -2420,16 +2419,16 @@ func Test_Closed_handlerAccountTransfer(t *testing.T) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
 					Admin:      keeper.AliceKey,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Type:       types.ClosedDomain,
 					Broker:     nil,
 				}).Create()
 				// account owned by bob
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
+					Name:       utils.StrPtr("test"),
 					Owner:      keeper.BobKey,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 				}).Create()
 			},
 			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
@@ -2455,7 +2454,7 @@ func Test_Closed_handlerAccountTransfer(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					panic("unexpected account deletion")
 				}
@@ -2470,16 +2469,16 @@ func Test_Closed_handlerAccountTransfer(t *testing.T) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
 					Admin:      keeper.AliceKey,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Type:       types.ClosedDomain,
 					Broker:     nil,
 				}).Create()
 				// account owned by bob
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("test"),
+					Name:         utils.StrPtr("test"),
 					Owner:        keeper.BobKey,
-					ValidUntil:   iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil:   utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					MetadataURI:  "lol",
 					Certificates: []types.Certificate{[]byte("test")},
 					Resources: []types.Resource{
@@ -2505,7 +2504,7 @@ func Test_Closed_handlerAccountTransfer(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					panic("unexpected account deletion")
 				}
@@ -2533,16 +2532,16 @@ func Test_Open_handlerAccountTransfer(t *testing.T) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
 					Admin:      keeper.AliceKey,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Type:       types.OpenDomain,
 					Broker:     nil,
 				}).Create()
 				// account owned by bob
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
+					Name:       utils.StrPtr("test"),
 					Owner:      keeper.BobKey,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 				}).Create()
 			},
 			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
@@ -2569,7 +2568,7 @@ func Test_Open_handlerAccountTransfer(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					panic("unexpected account deletion")
 				}
@@ -2584,16 +2583,16 @@ func Test_Open_handlerAccountTransfer(t *testing.T) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
 					Admin:      keeper.AliceKey,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Type:       types.OpenDomain,
 					Broker:     nil,
 				}).Create()
 				// account owned by bob
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("test"),
+					Name:         utils.StrPtr("test"),
 					Owner:        keeper.BobKey,
-					ValidUntil:   iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil:   utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					MetadataURI:  "lol",
 					Certificates: []types.Certificate{[]byte("test")},
 					Resources: []types.Resource{
@@ -2619,7 +2618,7 @@ func Test_Open_handlerAccountTransfer(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					t.Fatal("unexpected account deletion")
 				}
@@ -2684,8 +2683,8 @@ func Test_Common_handlerAccountTransfer(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					Admin:      iovns.ZeroAddress,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Admin:      keeper.AliceKey,
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Type:       types.OpenDomain,
 					Broker:     nil,
 				}).Create()
@@ -2707,14 +2706,14 @@ func Test_Common_handlerAccountTransfer(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					Admin:      iovns.ZeroAddress,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Admin:      keeper.BobKey,
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Type:       types.OpenDomain,
 					Broker:     nil,
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("test"),
+					Name:         utils.StrPtr("test"),
 					Owner:        keeper.BobKey,
 					ValidUntil:   0,
 					Resources:    nil,
@@ -2740,15 +2739,15 @@ func Test_Common_handlerAccountTransfer(t *testing.T) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
 					Admin:      keeper.AliceKey,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Type:       types.ClosedDomain,
 					Broker:     nil,
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("test"),
+					Name:         utils.StrPtr("test"),
 					Owner:        keeper.BobKey,
-					ValidUntil:   iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil:   utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Resources:    nil,
 					Certificates: nil,
 					Broker:       nil,
@@ -2771,16 +2770,16 @@ func Test_Common_handlerAccountTransfer(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					Admin:      iovns.ZeroAddress,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Admin:      keeper.BobKey,
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Type:       types.OpenDomain,
 					Broker:     nil,
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:       "test",
-					Name:         tutils.StrPtr("test"),
+					Name:         utils.StrPtr("test"),
 					Owner:        keeper.AliceKey,
-					ValidUntil:   iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil:   utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Resources:    nil,
 					Certificates: nil,
 					Broker:       nil,
@@ -2803,16 +2802,16 @@ func Test_Common_handlerAccountTransfer(t *testing.T) {
 			BeforeTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				executor.NewDomain(ctx, k, types.Domain{
 					Name:       "test",
-					Admin:      iovns.ZeroAddress,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					Admin:      keeper.BobKey,
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 					Type:       types.OpenDomain,
 					Broker:     nil,
 				}).Create()
 				executor.NewAccount(ctx, k, types.Account{
 					Domain:     "test",
-					Name:       tutils.StrPtr("test"),
+					Name:       utils.StrPtr("test"),
 					Owner:      keeper.AliceKey,
-					ValidUntil: iovns.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
+					ValidUntil: utils.TimeToSeconds(ctx.BlockTime().Add(1000 * time.Hour)),
 				}).Create()
 			},
 			Test: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
@@ -2828,7 +2827,7 @@ func Test_Common_handlerAccountTransfer(t *testing.T) {
 			},
 			AfterTest: func(t *testing.T, k keeper.Keeper, ctx sdk.Context, mocks *keeper.Mocks) {
 				account := new(types.Account)
-				exists := k.AccountStore(ctx).Read((&types.Account{Name: tutils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
+				exists := k.AccountStore(ctx).Read((&types.Account{Name: utils.StrPtr("test"), Domain: "test"}).PrimaryKey(), account)
 				if !exists {
 					panic("unexpected account deletion")
 				}